+++ packages/react/src/components/dropzone.tsx
@@ -66,6 +66,12 @@ export type UploadDropzoneProps<
    * @see https://docs.uploadthing.com/theming#content-customisation
    */
   content?: DropzoneContent;
+  /**
+   * Callback called when files are dropped or pasted.
+   *
+   * @param acceptedFiles - The files that were accepted.
+   */
+  onDrop?: (acceptedFiles: File[]) => void;
 };
 
 export function UploadDropzone<
@@ -116,7 +122,8 @@ export function UploadDropzone<
       headers: $props.headers,
       skipPolling: !$props?.onClientUploadComplete ? true : $props?.skipPolling,
       onClientUploadComplete: (res) => {
-        $props.onClientUploadComplete?.(res);
+        setFiles([]);
+        void $props.onClientUploadComplete?.(res);
         setUploadProgress(0);
       },
       onUploadProgress: (p, e) => {
@@ -132,11 +139,9 @@ export function UploadDropzone<
 
   const onDrop = useCallback(
     (acceptedFiles: File[]) => {
-      setFiles(
-        acceptedFiles.map((file) =>
-          Object.assign(file, { status: "pending" as const, key: null }),
-        ),
-      );
+      $props.onDrop?.(acceptedFiles);
+
+      setFiles(acceptedFiles);
 
       // If mode is auto, start upload immediately
       if (mode === "auto") {