+++ packages/solid/src/components/dropzone.tsx
@@ -62,6 +62,12 @@ export type UploadDropzoneProps<
    * @see https://docs.uploadthing.com/theming#content-customisation
    */
   content?: DropzoneContent;
+  /**
+   * Callback called when files are dropped or pasted.
+   *
+   * @param acceptedFiles - The files that were accepted.
+   */
+  onDrop?: (acceptedFiles: File[]) => void;
   config?: {
     mode?: "manual" | "auto";
   };
@@ -103,6 +109,8 @@ export const UploadDropzone = <
 
   const [files, setFiles] = createSignal<File[]>([]);
   const onDrop = (acceptedFiles: File[]) => {
+    $props.onDrop?.(acceptedFiles);
+
     setFiles(acceptedFiles);
 
     // If mode is auto, start upload immediately