+++ packages/uploadthing/src/internal/multi-part.browser.ts
@@ -9,7 +9,7 @@ import {
 } from "@uploadthing/shared";
 import type { ContentDisposition, UploadThingError } from "@uploadthing/shared";
 
-import type { MPUResponse } from "./types";
+import type { MPUResponse } from "./shared-schemas";
 import type { UTReporter } from "./ut-reporter";
 
 export const uploadMultipartWithProgress = (
@@ -47,7 +47,7 @@ export const uploadMultipartWithProgress = (
           Effect.retry({
             while: (error) => error instanceof RetryError,
             times: isTest ? 3 : 10, // less retries in tests just to make it faster
-            schedule: exponentialBackoff,
+            schedule: exponentialBackoff(),
           }),
         );
       },