+++ packages/uploadthing/src/types.ts
@@ -16,7 +16,10 @@ export type {
 
 export * from "./sdk/types";
 
-export type { EndpointMetadata } from "@uploadthing/shared";
+export type {
+  EndpointMetadata,
+  ExpandedRouteConfig,
+} from "@uploadthing/shared";
 
 export type {
   FileUploadData,