+++ packages/uploadthing/test/__test-helpers.ts
@@ -10,11 +10,11 @@ import { generateUploadThingURL } from "@uploadthing/shared";
 
 import { UPLOADTHING_VERSION } from "../src/internal/constants";
 import type {
-  ActionType,
   MPUResponse,
   PresignedBase,
   PSPResponse,
-} from "../src/internal/types";
+} from "../src/internal/shared-schemas";
+import type { ActionType } from "../src/internal/types";
 
 export const requestSpy = vi.fn<[string, RequestInit]>();
 export const requestsToDomain = (domain: string) =>
@@ -85,7 +85,7 @@ const mockPresigned = (file: {
 const callRequestSpy = async (request: StrictRequest<any>) =>
   requestSpy(new URL(request.url).toString(), {
     method: request.method,
-    headers: Object.fromEntries(request.headers.entries()),
+    headers: Object.fromEntries(request.headers),
     body: await (() => {
       if (request.method === "GET") return null;
       const ct = request.headers.get("content-type");
@@ -96,35 +96,11 @@ const callRequestSpy = async (request: StrictRequest<any>) =>
     })(),
   });
 
-const msw = setupServer(
-  /**
-   * S3
-   */
-  http.post("https://bucket.s3.amazonaws.com", async ({ request }) => {
-    await callRequestSpy(request);
-    return new HttpResponse();
-  }),
-  http.put("https://bucket.s3.amazonaws.com/:key", async ({ request }) => {
-    await callRequestSpy(request);
-    return new HttpResponse(null, {
-      status: 204,
-      headers: { ETag: "abc123" },
-    });
-  }),
-  /**
-   * Static Assets
-   */
-  http.get("https://cdn.foo.com/:fileKey", async ({ request }) => {
-    await callRequestSpy(request);
-    return HttpResponse.text("Lorem ipsum doler sit amet");
-  }),
-  http.get("https://utfs.io/f/:key", async ({ request }) => {
-    await callRequestSpy(request);
-    return HttpResponse.text("Lorem ipsum doler sit amet");
-  }),
-);
+const msw = setupServer();
 beforeAll(() => msw.listen({ onUnhandledRequest: "bypass" }));
 afterAll(() => msw.close());
+
+export const resetMocks = () => msw.close();
 
 /**
  * Extend the base `it` function to provide a `db` instance to our tests
@@ -142,8 +118,35 @@ export const it = itBase.extend({
       insertFile: (file: any) => files.push(file),
       getFileByKey: (key: string) => files.find((f) => f.key === key),
     };
-    // prepend msw listeners to use db instance
     msw.use(
+      /**
+       * S3
+       */
+      http.post("https://bucket.s3.amazonaws.com", async ({ request }) => {
+        await callRequestSpy(request);
+        return new HttpResponse();
+      }),
+      http.put("https://bucket.s3.amazonaws.com/:key", async ({ request }) => {
+        await callRequestSpy(request);
+        return new HttpResponse(null, {
+          status: 204,
+          headers: { ETag: "abc123" },
+        });
+      }),
+      /**
+       * Static Assets
+       */
+      http.get("https://cdn.foo.com/:fileKey", async ({ request }) => {
+        await callRequestSpy(request);
+        return HttpResponse.text("Lorem ipsum doler sit amet");
+      }),
+      http.get("https://utfs.io/f/:key", async ({ request }) => {
+        await callRequestSpy(request);
+        return HttpResponse.text("Lorem ipsum doler sit amet");
+      }),
+      /**
+       * UploadThing API
+       */
       http.post<never, { files: any[] } & Record<string, string>>(
         "https://uploadthing.com/api/prepareUpload",
         async ({ request }) => {