+++ packages/solid/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @uploadthing/solid
 
+## 6.4.4
+
+### Patch Changes
+
+- Updated dependencies [[`811b4cb`](https://github.com/pingdotgg/uploadthing/commit/811b4cb96938dd498f55e323f34685cbc8cfea9c), [`5e6e64c`](https://github.com/pingdotgg/uploadthing/commit/5e6e64c53ac9765ceee4bb758a48e08eabb36d14)]:
+  - @uploadthing/shared@6.7.4
+  - @uploadthing/dropzone@0.4.1
+
 ## 6.4.3
 
 ### Patch Changes